// src/models/RouteSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Stadia Maps Geospatial APIs
 * The Stadia Maps Geospatial APIs provide you with the data you need to build awesome applications. All endpoints are versioned individually to allow for granular upgrades. We follow the [Semantic Versioning scheme](https://semver.org/).
 *
 * The version of the OpenAPI document: 4.1.0
 * Contact: support@stadiamaps.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RouteSummary
 */
export interface RouteSummary {
    /**
     * The estimated travel time, in seconds
     * @type {number}
     * @memberof RouteSummary
     */
    time: number;
    /**
     * The estimated travel distance, in `units` (km or mi)
     * @type {number}
     * @memberof RouteSummary
     */
    length: number;
    /**
     * The minimum latitude of the bounding box containing the route.
     * @type {number}
     * @memberof RouteSummary
     */
    minLat: number;
    /**
     * The maximum latitude of the bounding box containing the route.
     * @type {number}
     * @memberof RouteSummary
     */
    maxLat: number;
    /**
     * The minimum longitude of the bounding box containing the route.
     * @type {number}
     * @memberof RouteSummary
     */
    minLon: number;
    /**
     * The maximum longitude of the bounding box containing the route.
     * @type {number}
     * @memberof RouteSummary
     */
    maxLon: number;
}

/**
 * Check if a given object implements the RouteSummary interface.
 */
export function instanceOfRouteSummary(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "time" in value;
    isInstance = isInstance && "length" in value;
    isInstance = isInstance && "minLat" in value;
    isInstance = isInstance && "maxLat" in value;
    isInstance = isInstance && "minLon" in value;
    isInstance = isInstance && "maxLon" in value;

    return isInstance;
}

export function RouteSummaryFromJSON(json: any): RouteSummary {
    return RouteSummaryFromJSONTyped(json, false);
}

export function RouteSummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): RouteSummary {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'time': json['time'],
        'length': json['length'],
        'minLat': json['min_lat'],
        'maxLat': json['max_lat'],
        'minLon': json['min_lon'],
        'maxLon': json['max_lon'],
    };
}

export function RouteSummaryToJSON(value?: RouteSummary | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'time': value.time,
        'length': value.length,
        'min_lat': value.minLat,
        'max_lat': value.maxLat,
        'min_lon': value.minLon,
        'max_lon': value.maxLon,
    };
}